refactor: move token related constants
