+++ src/frontend/src/lib/components/wallet/tokens/ReceiveTokensSigner.svelte
@@ -6,7 +6,8 @@
 	import Confetti from '$lib/components/ui/Confetti.svelte';
 	import Spinner from '$lib/components/ui/Spinner.svelte';
 	import ReceiveTokensSignerForm from '$lib/components/wallet/tokens/ReceiveTokensSignerForm.svelte';
-	import { ICP_TOKEN, OISY_WALLET_OPTIONS } from '$lib/constants/wallet.constants';
+	import { ICP } from '$lib/constants/token.constants';
+	import { OISY_WALLET_OPTIONS } from '$lib/constants/wallet.constants';
 	import type { WalletId } from '$lib/schemas/wallet.schema';
 	import { wizardBusy } from '$lib/stores/app/busy.store';
 	import { i18n } from '$lib/stores/app/i18n.store';
@@ -73,7 +74,7 @@
 			amount,
 			balance,
 			token: ICPToken,
-			fee: ICP_TOKEN.fee
+			fee: ICP.fee
 		});
 
 		if (!valid || isNullish(tokenAmount)) {
