+++ src/frontend/src/lib/services/wallet/wallet.approve.services.ts
@@ -1,9 +1,6 @@
 import { approveIcrcTransfer } from '$lib/api/icrc-ledger.api';
-import {
-	CONSOLE_CANISTER_ID,
-	CYCLES_LEDGER_CANISTER_ID,
-	CYCLES_TRANSACTION_FEE
-} from '$lib/constants/app.constants';
+import { CONSOLE_CANISTER_ID, CYCLES_LEDGER_CANISTER_ID } from '$lib/constants/app.constants';
+import { CYCLES_TRANSACTION_FEE } from '$lib/constants/token.constants';
 import { MEMO_CANISTER_APPROVE } from '$lib/constants/wallet.constants';
 import type { OptionIdentity } from '$lib/types/itentity';
 import type { IcpLedgerDid } from '@icp-sdk/canisters/ledger/icp';
